apps/backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-auditlog==3.4.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3